+++ tests/components/pulses/test_pulse_weights.py
@@ -24,20 +24,22 @@ def test_constant_readout_pulse_integration_weights_default():
     compare_integration_weights(expected_weights, weights)
 
 
-def test_empty_integration_weights():
-    for weights in ([], np.array([]), None):
-        pulse = pulses.SquareReadoutPulse(
-            length=100, amplitude=1, integration_weights=weights
-        )
+def test_default_integration_weights():
+    pulse = pulses.SquareReadoutPulse(length=100, amplitude=1)
+    assert pulse.default_integration_weights == [(1, 100)]
 
-        weights = pulse.integration_weights_function()
-        expected_weights = {
-            "real": [(1, 100)],
-            "imag": [(0, 100)],
-            "minus_real": [(-1, 100)],
-            "minus_imag": [(0, 100)],
-        }
-        compare_integration_weights(expected_weights, weights)
+
+def test_empty_integration_weights():
+    pulse = pulses.SquareReadoutPulse(length=100, amplitude=1)
+
+    weights = pulse.integration_weights_function()
+    expected_weights = {
+        "real": [(1, 100)],
+        "imag": [(0, 100)],
+        "minus_real": [(-1, 100)],
+        "minus_imag": [(0, 100)],
+    }
+    compare_integration_weights(expected_weights, weights)
 
 
 def test_constant_readout_pulse_integration_weights_phase_shift():