+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Changed
+- `Pulse.integration_weights` now defaults to `#./default_integration_weights`, which returns [(1, pulse.length)]
+
 ## [0.3.8]
 ### Added
 - Added time tagging to channels