+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## [0.3.9]
 ### Added
 - Added `QuamBase.set_at_reference` to set a value at a reference
 - Added `string_reference.get_parent_reference` to get the parent reference of a string reference